requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
dvc
dvc-s3
pyyaml
mlflow
dagshub